pyproject.toml
[build-system]
requires = ["setuptools>=68.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "modelforge"
version = "0.1.0"
description = "AI Model Sharing Center for Power Industry"
requires-python = ">=3.11"
dependencies = [
    "fastapi>=0.115,<1.0",
    "uvicorn[standard]>=0.27",
    "pydantic>=2.0",
    "pydantic-settings>=2.0",
    "python-multipart>=0.0.7",
    "pyyaml>=6.0",
    "pypinyin>=0.51",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0",
    "httpx>=0.27",
    "ruff>=0.3",
]
serving = [
    "scikit-learn>=1.4",
    "joblib>=1.3",
    "numpy>=1.26",
    "onnxruntime>=1.17",
]
example = [
    "scikit-learn>=1.4",
    "pandas>=2.1",
    "numpy>=1.26",
    "joblib>=1.3",
    "requests>=2.31",
]

[tool.setuptools.packages.find]
where = ["src"]

[tool.ruff]
target-version = "py312"
line-length = 100

[tool.ruff.lint]
select = ["E", "F", "I", "W"]

[tool.pytest.ini_options]
testpaths = ["tests"]
